Added heuristics & heuristic function
Each node in heuristics_dict now has the distance from node to target token and the nextNode to follow the shortest path. If path = 'np' it is not possible to reach the target token from current node.

getHeuristic() returns the next node that the path should follow to achieve the shortest path to target. If 'backtrack' is returned, calling function knows to go back up a level or check adjacent node.